Split logic into submodules
