+++ src/aind_behavior_video_transformation/__init__.py
@@ -1,3 +1,6 @@
 """Init package"""
 
 __version__ = "0.1.0"
+
+from .etl import BehaviorVideoJob, BehaviorVideoJobSettings  # noqa F401
+from .transform_videos import CompressionEnum, CompressionRequest  # noqa F401
